Complete storyboook stories.
This commit also fixes the issue of the `withComponent` method in the examples
which no longer appears to be valid styled-component api. Because the code for
implementing a button wrapper component relied on this api, it was refactored
to instead use the styled function wrapping a custom component with the styles
of the wrapped components defined separately from the component so that they
could be easily injected in the wrapper as well.